Scripts/es-data-loader/app/requirements.txt
elastic-transport==8.12.0 
elasticsearch==8.12.1
google-cloud-storage==2.14.0
